further skeleton

+++ src/lib.rs
@@ -42,7 +42,7 @@ pub fn isprime(i: usize) -> bool {
 
     return true;
 }
-pub fn printdivisors(i: u64) {
+pub fn printdivisors(i: usize) {
     for x in 2..(i / 2) {
         if i % x == 0 {
             print!("{x} ,");
